requirements.txt
psycopg2-binary
yoyo-migrations
pyramid
wsgicors
waitress